pyproject.toml
[project]
name = "uv"
version = "0.1.0"
description = "An asynchronous Python program that automates university course schedule retrieval."
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "asyncpg>=0.30.0",
    "fake-useragent==2.2.0",
    "line-bot-sdk>=3.19.0",
    "numpy==2.2.6",
    "opencv-contrib-python==4.10.0.84",
    "openpyxl>=3.1.5",
    "paddleocr==3.1.0",
    "paddlepaddle==3.1.0",
    "pandas>=2.3.1",
    "plotly>=6.3.0",
    "python-dotenv>=1.1.1",
    "pyyaml==6.0.2",
    "selenium==4.34.2",
    "setuptools>=80.9.0",
    "twilio>=9.8.0",
    "undetected-chromedriver==3.5.5",
]

[dependency-groups]
dev = [
    "ruff>=0.12.10",
]
